+++ src/main/kotlin/eu/jameshamilton/codegen/types.kt
@@ -9,14 +9,14 @@ sealed interface TypeX86 {
 
 data object Unknown : TypeX86 {
     override val size: Bytes
-        get() = 0
+        get() = 0L
     override val suffix: String
         get() = unreachable("Unknown suffix $this")
 }
 
 data object Byte_ : TypeX86 {
     override val size: Bytes
-        get() = 1
+        get() = 1L
     override val suffix: String
         get() = "b"
 
@@ -24,26 +24,26 @@ data object Byte_ : TypeX86 {
 
 data object Longword : TypeX86 {
     override val size: Bytes
-        get() = 4
+        get() = 4L
     override val suffix: String
         get() = "l"
 }
 
 data object Quadword : TypeX86 {
     override val size: Bytes
-        get() = 8
+        get() = 8L
     override val suffix: String
         get() = "q"
 }
 
 data object Double_ : TypeX86 {
     override val size: Bytes
-        get() = 8
+        get() = 8L
     override val suffix: String
         get() = "sd"
 }
 
-data class ByteArray(override val size: Int, val alignment: Int) : TypeX86 {
+data class ByteArray(override val size: Long, val alignment: Long) : TypeX86 {
 
     override val suffix: String
         get() = TODO()
